rename gymbutton to button
