--- src/components/GymButton/index.tsx
@@ -1,3 +1,0 @@
-import GymButton from './GymButton';
-
-export default GymButton;
