+++ src/index.ts
@@ -1,7 +1,6 @@
 import Button from './components/Button';
 import Link from './components/Link';
 import Card from './components/Card';
-import GymButton from './components/GymButton';
 import BulletedList from './components/BulletedList';
 import ListItem from './components/ListItem';
 import UserLogoutMenu from './components/UserLogoutMenu';
@@ -12,7 +11,6 @@ export {
   Button,
   Link,
   Card,
-  GymButton,
   BulletedList,
   ListItem,
   UserLogoutMenu,
